+++ app/models/facilities_management/rm6378/journey/procurement.rb
@@ -1,0 +1,7 @@
+module FacilitiesManagement
+  module RM6378
+    class Journey::Procurement
+      include Steppable
+    end
+  end
+end